+++ video.h
@@ -3,6 +3,6 @@
 #include "common.h"
 #include "SDL.h"
 
-const uint8_t *video_get_fb(void);
-void           video_draw_line(void);
-void           draw_screen(void);
+const uint16_t *video_get_fb(void);
+void            video_draw_line(void);
+void            draw_screen(void);
